backend/pyproject.toml
[project]
name = "balanceteshaters"
version = "0.1.0"
description = "Balance tes haters"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "colorlog>=6.10.1",
    "dependency-injector[pydantic2]>=4.48.2",
    "fastapi[standard]>=0.122.0",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "sqlalchemy[asyncio]>=2.0.44",
]
